+++ src/pages/register/index.ts
@@ -3,6 +3,7 @@ import { goToElementHref, stringifyProps } from "../../utils";
 import Wrapper from "../../components/wrapper";
 import Form, { Header, Footer, Body, Group, Label, Control } from "../../components/form";
 import Button from "../../components/button";
+import validator from "utils/validator";
 import template from "./index.tem";
 import "./index.css";
 
@@ -11,59 +12,11 @@ const loginRegExp = /^[a-zA-Z0-9.$_]{4,256}$/;
 const textRegExp = /^[a-zA-Zа-яА-Я.$_]{4,256}$/;
 const phoneRegExp = /^(\+\d|8)[ ()\d-]{10,16}$/;
 
-function valiateFormInput(element: HTMLInputElement) {
-  const group: HTMLElement = element.parentNode;
-  const value = element.value.trim();
-  const requireIsValid = element.required ? !!value : true;
 
-  group.classList[requireIsValid ? "remove" : "add"]("form__group_invalid-require");
-  if (!requireIsValid) return false;
-
-  switch (element.type) {
-    case "email": {
-      const emailIsValid = emailRegExp.test(value);
-
-      group.classList[emailIsValid ? "remove" : "add"]("form__group_invalid-email");
-      return emailIsValid;
-    }
-    case "text": {
-      if (element.name === "login") {
-        const loginIsValid = loginRegExp.test(value);
-
-        group.classList[loginIsValid ? "remove" : "add"]("form__group_invalid-login");
-        return loginIsValid;
-      }
-
-      const textIsValid = textRegExp.test(value);
-
-      group.classList[!requireIsValid || textIsValid ? "remove" : "add"]("form__group_invalid-text");
-      return textIsValid;
-    }
-    case "tel": {
-      const phoneStrIsValid = phoneRegExp.test(value);
-      const phoneNumberlength: number | null = value && value.match(/\d/g).length;
-
-      group.classList[phoneStrIsValid && phoneNumberlength === 11 ? "remove" : "add"]("form__group_invalid-phone");
-      return phoneStrIsValid && phoneNumberlength === 11;
-    }
-    case "password": {
-      if (element.name === "password2") {
-        const passIsValid = value === element.closest(".form").querySelector("[name='password']").value;
-
-        group.classList[!requireIsValid || passIsValid ? "remove" : "add"]("form__group_invalid-password");
-        return passIsValid;
-      }
-      break;
-    }
-    default:
-  }
-  return true;
-}
-
-function submit(event) {
+function submit(event: {target: HTMLButtonElement}) {
   const { target } = event;
-  const form = target.closest(".form");
-  const controls = form.querySelectorAll(".form__control");
+  const form: HTMLElement = target!.closest(".form")!;
+  const controls : HTMLInputElement[] = form.querySelectorAll(".form__control");
   let result = true;
   controls.forEach((el) => {
     result = valiateFormInput(el) || false;
@@ -71,9 +24,9 @@ function submit(event) {
   if (result) goToElementHref(event);
 }
 
-function validate(e) {
-  console.log(e.target);
-  valiateFormInput(e.target);
+function validate(event: Event) {
+  const {target} = event;
+  validator(target);
 }
 
 export default class Register extends Component {
